requirements.txt
streamlit
google-generativeai
radon
networkx
matplotlib
coverage
python-dotenv